fix the makefile
On other projects where we are using `BuildMethod: makefile`
we are just making a dependencies layer, so its ok for the
Makefile to just install packages.

On this project there is no dependency layer so if we
switch the `BuildMethod` to `makefile` the makefile is
also responsible for copying the application code into
the zip artifact.